Fix crash when user has not specified a timeout
In the case that the user has not set a timeout, a call to
Spreadsheet.rows will crash with no function clause matching in :gen.call/4

This is because :gen.call with a fourth argument expects it to be an
integer, not nil.

We can solve this by using the default if the user did not configure a
timeout.